+++ packages/libs/lambda-at-edge/tests/default-handler/default-handler-with-404.test.ts
@@ -1,6 +1,6 @@
 import { handler } from "../../src/default-handler";
 import { createCloudFrontEvent } from "../test-utils";
-import { CloudFrontResultResponse } from "aws-lambda";
+import { CloudFrontRequest, CloudFrontResultResponse } from "aws-lambda";
 
 jest.mock(
   "../../src/manifest.json",
@@ -39,8 +39,45 @@ describe("Lambda@Edge", () => {
       config: { eventType: "origin-request" } as any
     });
 
+    const request = (await handler(event)) as CloudFrontRequest;
+
+    expect(request.uri).toEqual("/404.html");
+  });
+
+  it.each`
+    path
+    ${"/_next/data/unmatched"}
+  `(
+    "renders a static 404 page if data request can't be matched for path: $path",
+    async ({ path }) => {
+      const event = createCloudFrontEvent({
+        uri: path,
+        origin: {
+          s3: {
+            domainName: "my-bucket.s3.amazonaws.com"
+          }
+        },
+        config: { eventType: "origin-request" } as any
+      });
+
+      const request = (await handler(event)) as CloudFrontRequest;
+
+      expect(request.uri).toEqual("/404.html");
+    }
+  );
+
+  it("static 404 page should return CloudFront 404 status code after successful S3 origin response", async () => {
+    const event = createCloudFrontEvent({
+      uri: "/404.html",
+      host: "mydistribution.cloudfront.net",
+      config: { eventType: "origin-response" } as any,
+      response: {
+        status: "200"
+      } as any
+    });
+
     const response = (await handler(event)) as CloudFrontResultResponse;
 
-    expect(response.uri).toEqual("/404.html");
+    expect(response.status).toEqual("404");
   });
 });