+++ packages/libs/lambda-at-edge/tests/integration/with-trailing-slash-next-config/fixture-next-config-as-obj-basepath/next.config.js
@@ -1,0 +1,1 @@
+module.exports = { basePath: "/basepath", target: "serverless", trailingSlash: true };