+++ packages/libs/lambda-at-edge/tests/shared-fixtures/built-artifact/pages/erroredPage.js
@@ -1,0 +1,5 @@
+module.exports = {
+  render: (req, res) => {
+    throw Error("This page had an error while rendering!");
+  }
+};