+++ packages/libs/lambda-at-edge/tests/shared-fixtures/built-artifact/pages/preview.js
@@ -1,0 +1,14 @@
+module.exports = {
+  render: (req, res) => {
+    res.end("pages/preview.js");
+  },
+  renderReqToHTML: (req, res) => {
+    return {
+      renderOpts: {
+        pageData: {
+          page: "pages/preview.js"
+        }
+      }
+    };
+  }
+};