+++ packages/serverless-components/aws-lambda/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.3-alpha.0](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/aws-lambda@1.0.2...@sls-next/aws-lambda@1.0.3-alpha.0) (2020-10-07)
+
+**Note:** Version bump only for package @sls-next/aws-lambda
+
+## [1.0.2](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/aws-lambda@1.0.2-alpha.1...@sls-next/aws-lambda@1.0.2) (2020-09-26)
+
+**Note:** Version bump only for package @sls-next/aws-lambda
+
 ## [1.0.2-alpha.1](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/aws-lambda@1.0.2-alpha.0...@sls-next/aws-lambda@1.0.2-alpha.1) (2020-08-14)
 
 **Note:** Version bump only for package @sls-next/aws-lambda
@@ -43,11 +51,11 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 **Note:** Version bump only for package @sls-next/aws-lambda
 
-# [1.0.0-alpha.2](https://github.com/danielcondemarin/serverless-next.js/compare/@sls-next/aws-lambda@1.0.0-alpha.1...@sls-next/aws-lambda@1.0.0-alpha.2) (2020-06-24)
+# [1.0.0-alpha.2](https://github.com/serverless-nextjs/serverless-next.js/compare/@sls-next/aws-lambda@1.0.0-alpha.1...@sls-next/aws-lambda@1.0.0-alpha.2) (2020-06-24)
 
 ### Bug Fixes
 
-- **serverless-component:** don't overwrite the cloudfront default.forward config ([#460](https://github.com/danielcondemarin/serverless-next.js/issues/460)) ([12da1de](https://github.com/danielcondemarin/serverless-next.js/commit/12da1de31855b68b9addef801ec21dffd3202a21))
+- **serverless-component:** don't overwrite the cloudfront default.forward config ([#460](https://github.com/serverless-nextjs/serverless-next.js/issues/460)) ([12da1de](https://github.com/serverless-nextjs/serverless-next.js/commit/12da1de31855b68b9addef801ec21dffd3202a21))
 
 # 1.0.0-alpha.1 (2020-06-21)